Proposed Change to Spending Legendary Points
The ability to apply Legend Points after the result goes against the "every roll matters" mantra to me and seems a bit min-max gimmicky to me. With the previous rules, you would never want to apply the points prior to the roll as you could potentially waste them with a high initial roll. I feel like changing this to only working prior to the roll but giving a little bit added benefit makes the decision of spending them more important.